Day 17 - Part 2 - Back to long for A

+++ src/main/kotlin/2024/Day17.kt
@@ -7,7 +7,7 @@ import java.time.LocalDateTime
 import kotlin.math.pow
 
 data class Registers(
-    val a: Int,
+    val a: Long,
     val b: Int,
     val c: Int,
     val output: List<Int> = emptyList(),
@@ -16,15 +16,15 @@ data class Registers(
     fun updateRegister(opcode: Int, operand: Int): Registers {
         val comboOperand = when (operand) {
             4 -> a
-            5 -> b
-            6 -> c
-            else -> operand
+            5 -> b.toLong()
+            6 -> c.toLong()
+            else -> operand.toLong()
         }
         return when (opcode) {
-            0 -> copy(a = a.div(2.toDouble().pow(comboOperand.toDouble())).toInt())
+            0 -> copy(a = a.div(2.toDouble().pow(comboOperand.toDouble())).toLong())
             1 -> copy(b = b.xor(operand))
             2 -> copy(b = comboOperand.mod(8))
-            3 -> if (a == 0) this else copy(nextInstruction = operand)
+            3 -> if (a == 0L) this else copy(nextInstruction = operand)
             4 -> copy(b = b.xor(c))
             5 -> copy(output = output + comboOperand.mod(8))
 //            6 -> copy(b = a.div(2.toDouble().pow(comboOperand.toDouble())).toLong())
@@ -65,13 +65,13 @@ fun main() {
     val start = System.currentTimeMillis()
 
     fun mapProgramInstructionsAndRegister(input: List<String>): Pair<List<Int>, Registers> {
-        var a = 0
+        var a = 0L
         var b = 0
         var c = 0
         var programInstructions = listOf<Int>()
         input.forEach { line ->
             when {
-                line.contains("Register A") -> a = line.substringAfter(":").trim().toInt()
+                line.contains("Register A") -> a = line.substringAfter(":").trim().toLong()
                 line.contains("Register B") -> b = line.substringAfter(":").trim().toInt()
                 line.contains("Register C") -> c = line.substringAfter(":").trim().toInt()
                 line.contains("Program") -> {
@@ -90,16 +90,16 @@ fun main() {
         return runProgram(registers, programInstructions).output.joinToString(",")
     }
 
-    fun part2(input: List<String>): Int {
+    fun part2(input: List<String>): Long {
         val (programInstructions, registers) = mapProgramInstructionsAndRegister(input)
-        var a = 0
+        var a = 0L
         val dec = DecimalFormat("#,###.##")
         while (true) {
             val output = runProgram(registers.copy(a = a), programInstructions, true).output
             if (output == programInstructions) {
                 return a.also { "Found the answer!".println() }
             } else {
-                if (a % 10_000_000 == 0) "${LocalDateTime.now()} - A is ${dec.format(a)}.".println()
+                if (a % 10_000_000 == 0L) "${LocalDateTime.now()} - A is ${dec.format(a)}.".println()
                 a++
             }
         }
@@ -124,7 +124,7 @@ fun main() {
 
         Program: 0,3,5,4,3,0
     """.trimIndent().lines()
-    check(part2(sampleInputPart2) == 117440)
+    check(part2(sampleInputPart2) == 117440L)
     part2(input).println()
 
     "${(System.currentTimeMillis() - start)} milliseconds".println()
